+++ statistics-sdk/src/main/java/com/dwarfeng/statistics/sdk/bean/entity/FastJsonTask.java
@@ -16,7 +16,7 @@ import java.util.Objects;
  */
 public class FastJsonTask implements Bean {
 
-    private static final long serialVersionUID = -3048730124737064709L;
+    private static final long serialVersionUID = -7589362386673546983L;
 
     public static FastJsonTask of(Task task) {
         if (Objects.isNull(task)) {
@@ -31,8 +31,7 @@ public class FastJsonTask implements Bean {
                     task.getShouldExpireDate(),
                     task.getShouldDieDate(),
                     task.getExecutionNodeId(),
-                    task.getMessage(),
-                    task.getRemark()
+                    task.getMessage()
             );
         }
     }
@@ -64,15 +63,12 @@ public class FastJsonTask implements Bean {
     @JSONField(name = "message", ordinal = 9)
     private String message;
 
-    @JSONField(name = "remark", ordinal = 10)
-    private String remark;
-
     public FastJsonTask() {
     }
 
     public FastJsonTask(
             FastJsonLongIdKey key, FastJsonLongIdKey statisticsSettingKey, int status, Date createDate, Date startDate,
-            Date shouldExpireDate, Date shouldDieDate, int executionNodeId, String message, String remark
+            Date shouldExpireDate, Date shouldDieDate, int executionNodeId, String message
     ) {
         this.key = key;
         this.statisticsSettingKey = statisticsSettingKey;
@@ -83,7 +79,6 @@ public class FastJsonTask implements Bean {
         this.shouldDieDate = shouldDieDate;
         this.executionNodeId = executionNodeId;
         this.message = message;
-        this.remark = remark;
     }
 
     public FastJsonLongIdKey getKey() {
@@ -158,14 +153,6 @@ public class FastJsonTask implements Bean {
         this.message = message;
     }
 
-    public String getRemark() {
-        return remark;
-    }
-
-    public void setRemark(String remark) {
-        this.remark = remark;
-    }
-
     @Override
     public String toString() {
         return "FastJsonTask{" +
@@ -178,7 +165,6 @@ public class FastJsonTask implements Bean {
                 ", shouldDieDate=" + shouldDieDate +
                 ", executionNodeId=" + executionNodeId +
                 ", message='" + message + '\'' +
-                ", remark='" + remark + '\'' +
                 '}';
     }
 }